no commit message given

+++ plugins/minikube.bash
@@ -14,10 +14,21 @@ function _install_minikube() {
 	# https://minikube.sigs.k8s.io/docs/start/
 	folder="${HOME}/install/minikube"
 	exec="${folder}/minikube"
-	rm -rf "${folder}"
+	rm -rf "${folder}" || true
 	mkdir -p "${folder}"
 	curl --location --silent --output "${exec}" "https://storage.googleapis.com/minikube/releases/latest/minikube-linux-amd64"
 	chmod +x "${exec}"
+}
+
+function _uninstall_minikube() {
+	folder="${HOME}/install/minikube"
+	if [ -f "${folder}" ]
+	then
+		echo "remoing minikube folder"
+		rm -rf "${folder}"
+	else
+		echo "no minikube detected"
+	fi
 }
 
 register _activate_minikube
